Move root analyzer tests to brokk-shared
